Refactor CommunityTransactions to use CommunityOnchainTransactions, removing the deprecated CommunityTransactions component and its related files for improved clarity and maintainability.
