+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/AdminContestsPage/AdminContestsPage.tsx
@@ -22,9 +22,9 @@ import { CWButton } from 'views/components/component_kit/new_designs/CWButton';
 import CWPageLayout from 'views/components/component_kit/new_designs/CWPageLayout';
 import { PageNotFound } from 'views/pages/404';
 import EmptyCard from 'views/pages/CommunityManagement/Contests/EmptyContestsList/EmptyCard';
-import CommunityTransactions from 'views/pages/CreateCommunity/steps/CommunityTransactions';
 
 import { CWDivider } from '../../../../components/component_kit/cw_divider';
+import CommunityOnchainTransactions from '../../../CreateCommunity/steps/CommunityOnchainTransactions';
 import ContestsList from '../ContestsList';
 import EmptyContestsList from '../EmptyContestsList';
 import { ContestType, ContestView } from '../types';
@@ -262,7 +262,7 @@ Set one up first.`
             />
           </div>
         ) : contestView === ContestView.NamespaceEnablemenement ? (
-          <CommunityTransactions
+          <CommunityOnchainTransactions
             createdCommunityName={community?.name}
             createdCommunityId={community?.id || ''}
             selectedAddress={selectedAddress!}
