+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Topics/Topics.tsx
@@ -5,7 +5,7 @@ import app from 'state';
 import CWFormSteps from 'views/components/component_kit/new_designs/CWFormSteps';
 import CWPageLayout from 'views/components/component_kit/new_designs/CWPageLayout';
 import StakeIntegration from 'views/pages/CommunityManagement/StakeIntegration';
-import CommunityTransactions from 'views/pages/CreateCommunity/steps/CommunityTransactions';
+import CommunityOnchainTransactions from 'views/pages/CreateCommunity/steps/CommunityOnchainTransactions';
 
 import TopicDetails from './TopicDetails';
 import WVConsent from './WVConsent';
@@ -154,7 +154,7 @@ export const Topics = () => {
         );
       case CreateTopicStep.WVNamespaceEnablement:
         return (
-          <CommunityTransactions
+          <CommunityOnchainTransactions
             createdCommunityName={community?.name}
             createdCommunityId={community?.id || ''}
             selectedAddress={selectedAddress!}
