+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityOnchainTransactions/EnableStake/EnableStake.scss
