+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityOnchainTransactions/SignCommunityTransactions/SignCommunityTransactions.tsx
