+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityOnchainTransactions/SignStakeTransactions/SignStakeTransactions.tsx
