+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityOnchainTransactions/SignStakeTransactions/useReserveCommunityNamespace.tsx
@@ -11,12 +11,9 @@ import {
 } from 'shared/analytics/types';
 import { useUpdateCommunityMutation } from 'state/api/communities';
 import useUserStore from 'state/ui/user';
+import { useNamespaceFactory } from '..';
 import useAppStatus from '../../../../../../hooks/useAppStatus';
-import { useNamespaceFactory } from '../../CommunityTransactions';
-import {
-  ActionState,
-  defaultActionState,
-} from '../../CommunityTransactions/types';
+import { ActionState, defaultActionState } from '../types';
 interface UseReserveCommunityNamespaceProps {
   communityId: string;
   namespace: string;
