+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityOnchainTransactions/helpers/index.ts
